Adding support for automate actions for API Client
- Expose MiqAeDomain's as /api/automate_domains
- Keep /api/automate for automate browser
- Move refresh_from_source action under /api/automate_domains
- Move refresh_from_source tests under automate_domains_spec.rb
- Adding rspecs

Fixes #594
